+++ OpenCoverTest.App/App.xaml.cs
@@ -14,27 +14,4 @@ namespace OpenCoverTest.App
     public partial class App : Application
     {
     }
-
-    public class Calculator
-    {
-        public int Add(int a, int b) => a + b;
-        public int Subtract(int a, int b) => a - b;
-    }
-
-
-
-
-    public class StringHelper
-    {
-        public string Reverse(string input)
-        {
-            if (input == null) return null;
-            return new string(input.Reverse().ToArray());
-        }
-    }
-
-    public class MathUtils
-    {
-        public bool IsEven(int number) => number % 2 == 0;
-    }
 }
