+++ OpenCoverTest.Tests/OpenCoverTest.Tests.csproj
@@ -116,9 +116,9 @@
     <Analyzer Include="..\packages\xunit.analyzers.1.18.0\analyzers\dotnet\cs\xunit.analyzers.fixes.dll" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\OpenCoverTest.App\OpenCoverTest.csproj">
-      <Project>{e4862578-7159-4d1e-a742-4a4b18b21daa}</Project>
-      <Name>OpenCoverTest</Name>
+    <ProjectReference Include="..\OpenCoverTest.Core\OpenCoverTest.Core.csproj">
+      <Project>{7e49c6be-0276-4115-8ff4-fa20f4e67a07}</Project>
+      <Name>OpenCoverTest.Core</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
