+++ OpenCoverTest.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenCoverTest", "OpenCoverT
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenCoverTest.Tests", "OpenCoverTest.Tests\OpenCoverTest.Tests.csproj", "{B13B8C41-D17E-485B-A45B-86D948AB725E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenCoverTest.Core", "OpenCoverTest.Core\OpenCoverTest.Core.csproj", "{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,14 @@ Global
 		{B13B8C41-D17E-485B-A45B-86D948AB725E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B13B8C41-D17E-485B-A45B-86D948AB725E}.Release|x64.ActiveCfg = Release|x64
 		{B13B8C41-D17E-485B-A45B-86D948AB725E}.Release|x64.Build.0 = Release|x64
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Debug|x64.Build.0 = Debug|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Release|x64.ActiveCfg = Release|Any CPU
+		{7E49C6BE-0276-4115-8FF4-FA20F4E67A07}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
